코드 리팩토링 작업중

+++ app/src/main/java/org/jeonfeel/moeuibit2/ui/coindetail/order/UpbitCoinOrder.kt
@@ -131,9 +131,6 @@ class UpbitCoinOrder @Inject constructor(private val upbitCoinOrderUseCase: Upbi
         }
     }
 
-    /**
-     *
-     */
     suspend fun requestBid(
         market: String,
         totalPrice: Double,
@@ -160,7 +157,7 @@ class UpbitCoinOrder @Inject constructor(private val upbitCoinOrderUseCase: Upbi
         } else {
             upbitCoinOrderUseCase.requestBTCBid(
                 market = market,
-                totalPrice = totalPrice.toDouble(),
+                totalPrice = totalPrice,
                 coin = MyCoin(
                     market = market,
                     purchasePrice = coinPrice.toDouble(),
